# cloud-brain/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zuralog Cloud Brain",
    "description": "AI Health Assistant Backend \u2014 Cross-app reasoning and autonomous actions.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "description": "Register a new user via Supabase Auth.\n\nCreates the user in Supabase and syncs them to the local `users`\ntable. Returns session tokens for immediate use.\n\nArgs:\n    body: Email and password for registration.\n    auth_service: Injected Supabase auth service.\n    db: Injected async database session.\n\nReturns:\n    AuthResponse with user_id, access_token, refresh_token, expires_in.\n\nRaises:\n    HTTPException: 400 if registration fails (e.g., email already taken).",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Log in an existing user via Supabase Auth.\n\nAuthenticates the user with Supabase and syncs them to the local\n`users` table (idempotent upsert). Returns session tokens.\n\nArgs:\n    body: Email and password for login.\n    auth_service: Injected Supabase auth service.\n    db: Injected async database session.\n\nReturns:\n    AuthResponse with user_id, access_token, refresh_token, expires_in.\n\nRaises:\n    HTTPException: 401 if credentials are invalid.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "description": "Log out the current user by invalidating their Supabase session.\n\nUses the user's access token (from the Authorization header)\nto call Supabase logout \u2014 NOT the service key.\n\nArgs:\n    credentials: Bearer token from the Authorization header.\n    auth_service: Injected Supabase auth service.\n\nReturns:\n    MessageResponse confirming logout.\n\nRaises:\n    HTTPException: 400 if logout fails.",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh",
        "description": "Exchange a refresh token for a new session.\n\nArgs:\n    body: The refresh token from a prior login.\n    auth_service: Injected Supabase auth service.\n\nReturns:\n    AuthResponse with new access_token, refresh_token, and expires_in.\n\nRaises:\n    HTTPException: 401 if the refresh token is invalid or expired.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/me/preferences": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Preferences",
        "description": "Get the current user's AI preferences.\n\nArgs:\n    credentials: Bearer token from the Authorization header.\n    auth_service: Injected auth service for token validation.\n    db: Injected async database session.\n\nReturns:\n    A dict with coach_persona, subscription_tier, and is_premium fields.\n\nRaises:\n    HTTPException: 404 if the user is not found in the database.",
        "operationId": "get_preferences_api_v1_users_me_preferences_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Preferences Api V1 Users Me Preferences Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Update Preferences",
        "description": "Update the current user's AI preferences.\n\nArgs:\n    body: The request body containing fields to update.\n    credentials: Bearer token from the Authorization header.\n    auth_service: Injected auth service for token validation.\n    db: Injected async database session.\n\nReturns:\n    A confirmation dict with the updated persona.\n\nRaises:\n    HTTPException: 400 if the persona is invalid or no fields provided.",
        "operationId": "update_preferences_api_v1_users_me_preferences_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Update Preferences Api V1 Users Me Preferences Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "AuthResponse",
        "description": "Successful authentication response.\n\nReturned on register, login, and refresh.\n\nAttributes:\n    user_id: Supabase UID of the authenticated user.\n    access_token: Short-lived JWT for API authorization.\n    refresh_token: Long-lived token for silent session renewal.\n    expires_in: Token lifetime in seconds."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "description": "Credentials for user login.\n\nAttributes:\n    email: User's email address.\n    password: User's password (min 6 chars enforced by Supabase)."
      },
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "description": "Simple message response.\n\nAttributes:\n    message: Human-readable status message."
      },
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "description": "Request body for token refresh.\n\nAttributes:\n    refresh_token: The long-lived refresh token from a prior login."
      },
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "description": "Credentials for new user registration.\n\nAttributes:\n    email: User's email address.\n    password: User's password (min 6 chars enforced by Supabase)."
      },
...
      "UpdatePreferencesRequest": {
        "properties": {
          "coach_persona": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coach Persona"
          }
        },
        "type": "object",
        "title": "UpdatePreferencesRequest",
        "description": "Request body for updating user preferences.\n\nAttributes:\n    coach_persona: The coaching style preference."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}